+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * Fix hookpatching a function that starts with CALL instruction
 * Update hookpatch method
 * Fix no-recording backup movie not stored in savestates
+* [ImGui] Force redraw before state saving/loading
 
 ## [1.4.5] - 2023-10-22
 ### Added